requirements.txt
alembic==1.16.5
Brotli==1.1.0
brotlicffi==1.1.0.0
cachetools==5.3.0
celery==5.5.3
chardet==5.2.0
colorama==0.4.6
cryptography==45.0.7
docutils==0.20
fastapi==0.116.1
filelock==3.19.1
groq==0.31.0
gunicorn
h2==4.3.0
httpx==0.28.1
ipython==8.12.3
ipywidgets==8.1.7
pyjnius==1.6.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.4.0
keyring==25.6.0
langgraph==0.6.0
nest-asyncio==1.6.0
Pillow==11.3.0
prometheus_client==0.22.1
protobuf==6.32.0
circuitbreaker==2.1.3
pydantic==2.11.7
pydantic_settings==2.10.1
pyOpenSSL==25.1.0
redis==6.4.0
simplejson==3.20.1
Sphinx==8.2.3
SQLAlchemy==2.0.15
structlog==25.4.0
urllib3_secure_extra==0.1.0
uvicorn==0.35.0
zstandard==0.23.0
python-dotenv==1.0.1